Speed up fixed PatchMatch SSD (dab -10%, stroke -3%)
45-case corpus: zero MAE, PSNR, seam, or texture differences.
Full kernel: 30-33 ns -> 16-18 ns; bounded half-cost: 13 ns.

+++ internal/patchmatch/cost_opaque_fallback.go
@@ -8,6 +8,6 @@ func synthesisEncryptCounterHardware(_ *[16]byte, _ *[11][16]byte) {}
 
 func pmOpaqueKernelAvailable() bool { return false }
 
-func pmRunSynthesisOpaqueKernel(args *pmOpaqueKernelArgs) float32 {
+func pmRunSynthesisOpaqueKernel(args *pmOpaqueKernelArgs) uint32 {
 	panic("opaque PatchMatch kernel is unavailable on this architecture")
 }
